QoL updates, tmux support, tasks support

+++ src/commands/cluster/delete.rs
@@ -1,9 +1,10 @@
 use crate::database::models::{Cluster, ClusterState};
+use crate::utils;
 
 use anyhow::Result;
 use sqlx::sqlite::SqlitePool;
 
-pub async fn delete(pool: &SqlitePool, cluster_id: &str) -> Result<()> {
+pub async fn delete(pool: &SqlitePool, cluster_id: &str, skip_confirmation: bool) -> Result<()> {
     let cluster = match Cluster::fetch_by_id(pool, cluster_id).await? {
         Some(cluster) => cluster,
         None => {
@@ -26,6 +27,13 @@ pub async fn delete(pool: &SqlitePool, cluster_id: &str) -> Result<()> {
                 cluster.state
             );
         }
+    }
+
+    if !(utils::user_confirmation(
+        skip_confirmation,
+        "Do you confirm you want to delete this cluster from the database?",
+    )?) {
+        return Ok(());
     }
 
     Cluster::delete(pool, cluster_id).await?;
